workers/seo/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0
